requirements.txt
pandas==2.1.4
numpy==1.24.3
openpyxl==3.1.2
matplotlib==3.7.2
seaborn==0.12.2
pymodbus>=3.0.0  
mysql-connector-python>=8.0.0  
PyYAML>=6.0